Phase A9: extract inline #[cfg(test)] mod tests blocks to sibling files
All 9 inline test blocks extracted to <module>/tests.rs siblings:
graphql, registry, identity, error, retry, completion_factory, mcp_pool,
admission/stream_guard, toolset/shared/context. No logic changes.
